+++ lib/WORKER_SANAYEE/State_warshe_accept.dart
@@ -986,7 +986,7 @@ class  _State_warshe_accept extends State<State_warshe_accept> {
         context: context,
         builder: (BuildContext context) {
           return Directionality(textDirection: ui.TextDirection.rtl,
-            child:delete_order(chooseDate:widget.ChooseDate,phoneuser:widget.phoneuser,phoneworker:widget.phoneworker,image:widget.image,description:widget.description,datesend:widget.datesend,timesend:widget.timesend,id:widget.id,phone:widget.phoneworker,date:widget.date,namefirst: widget.namefirst,namelast:widget.namelast,time:widget.time,),);
+            child:delete_order(work:widget.work,Information:widget.Information,Experiance:widget.Experiance,token:widget.tokenworker,lat:widget.lat,lng:widget.lng,name:widget.workername,chooseDate:widget.ChooseDate,phoneuser:widget.phoneuser,phoneworker:widget.phoneworker,image:widget.image,description:widget.description,datesend:widget.datesend,timesend:widget.timesend,id:widget.id,phone:widget.phoneworker,date:widget.date,namefirst: widget.namefirst,namelast:widget.namelast,time:widget.time,),);
         });
   }
 }
@@ -1109,7 +1109,7 @@ class  _accept extends State<accept> {
                     //   ),
                     //   ),
                     // ),
-                    Container(
+                    widget.description!=null?Container(
                       width:360,
                       height: 60,
                       color:Colors.white,
@@ -1121,7 +1121,7 @@ class  _accept extends State<accept> {
                         fontFamily: 'Changa',
                       ),
                       ),
-                    ),
+                    ):Container(height: 30,),
                   ],
                 ),
               ),
@@ -1182,7 +1182,7 @@ class  _accept extends State<accept> {
                   ],
                 ),
               ),
-              image!=null && !press?Container(
+              widget.orderimage!=null && !press?Container(
                 // print(_image[index].id+"");
                 width: 200,
                 height: 150,
@@ -1197,6 +1197,7 @@ class  _accept extends State<accept> {
                 ),
               ):Container(),
               Container(
+                height: 40,
                 // transform: Matrix4.translationValues(0.0, -25, 0.0),
                 margin: EdgeInsets.only(top: 0,right:10),
                 child:Row(
@@ -1205,12 +1206,12 @@ class  _accept extends State<accept> {
                     !press?GestureDetector(
                       onTap: (){
                         press=!press;
-                        h=200;
+                        h=180;
                         setState(() {
                         });
                       },
                       child: Container(
-                        margin: EdgeInsets.only(top: 70,),
+                        margin: EdgeInsets.only(top: 0,),
                         child: Row(
                           children: [
                             // Text('عرض أقل',
@@ -1228,7 +1229,7 @@ class  _accept extends State<accept> {
                     GestureDetector(
                       onTap: (){
                         press=!press;
-                        if(image==null)h=190;
+                        if(widget.orderimage==null)h=180;
                         else h=370;
                         setState(() {
 
@@ -1256,7 +1257,7 @@ class  _accept extends State<accept> {
               Container(
                 height: 300,
                 width: 450,
-                margin: press?EdgeInsets.only(top:10):EdgeInsets.only(top:30),
+                margin: press && widget.orderimage==null?EdgeInsets.only(top:10):EdgeInsets.only(top:10),
                 padding:EdgeInsets.only(right:0,left: 0),
                 decoration: BoxDecoration(
 
@@ -1559,7 +1560,7 @@ class  _accept extends State<accept> {
         context: context,
         builder: (BuildContext context) {
           return Directionality(textDirection: ui.TextDirection.rtl,
-            child:delete_order(chooseDate:widget.ChooseDate,phoneuser:widget.phoneuser,phoneworker:widget.phoneworker,image:widget.image,description:widget.description,datesend:widget.datesend,timesend:widget.timesend,id:widget.id,phone:widget.phoneworker,date:widget.date,namefirst: widget.namefirst,namelast:widget.namelast,time:widget.time,),);
+            child:delete_order(work:widget.work,Information:widget.Information,Experiance:widget.Experiance,token:widget.token,lat:widget.lat,lng:widget.lng,name:widget.workername,chooseDate:widget.ChooseDate,phoneuser:widget.phoneuser,phoneworker:widget.phoneworker,image:widget.image,description:widget.description,datesend:widget.datesend,timesend:widget.timesend,id:widget.id,phone:widget.phoneworker,date:widget.date,namefirst: widget.namefirst,namelast:widget.namelast,time:widget.time,),);
         });
   }
 }
@@ -1584,7 +1585,13 @@ class delete_order extends StatefulWidget {
   final timeend;
   final chooseDate;
   final name;
-  delete_order({this.name,this.chooseDate,this.timeend,this.timestart,this.id,this.phone,this.time,this.date,this.phoneworker,this.phoneuser,this.timesend,this.datesend,this.namelast,this.namefirst,this.description,this.image,this.country});
+  final lat;
+  final lng;
+  final Information;
+  final Experiance;
+  final work;
+  final token;
+  delete_order({this.token,this.lat,this.lng,this.Information,this.Experiance,this.work,this.name,this.chooseDate,this.timeend,this.timestart,this.id,this.phone,this.time,this.date,this.phoneworker,this.phoneuser,this.timesend,this.datesend,this.namelast,this.namefirst,this.description,this.image,this.country});
   _delete_order createState() => new _delete_order();
 
 }
@@ -1614,9 +1621,7 @@ class _delete_order extends State<delete_order> {
                     GestureDetector(
                       onTap: () async{
                         await delete_Order();
-                        print('ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff');
-                        DateTime date =DateTime.now();
-                        Navigator.push(context, MaterialPageRoute(builder: (context) => accept_order(time: widget.chooseDate,phone: widget.phoneworker,name: widget.name,),),);
+                        Navigator.push(context, MaterialPageRoute(builder: (BuildContext context) => Warshat(lat:widget.lat,lng:widget.lng,Information:widget.Information,Experiance:widget.Experiance,Work:widget.work,namelast:widget.namelast,image:widget.image,token:widget.token,namefirst:widget.namefirst,name:widget.name,phone: widget.phoneworker)));
                         print('=========================================================================');
 
                         // Navigator.push(context, MaterialPageRoute(builder: (context) => user_reserve_order(phoneuser:widget.phoneuser,username:widget.name,),),);
@@ -1663,16 +1668,13 @@ class _delete_order extends State<delete_order> {
   }
 
   Future delete_Order() async {
-    print(widget.id);
     DateTime date=DateTime.now();
+    print(widget.id);
     var formattedDate = DateFormat('yyyy-MM-dd').format(date);
     var formattedTime = DateFormat('HH:mm:ss').format(date);
-    var url = 'https://' + IP4 + '/testlocalhost/delete_accept_order.php';
+    var url = 'https://' + IP4 + '/testlocalhost/delete_warsha.php';
     var ressponse = await http.post(url, body: {
       "id": widget.id,
-      "datecancel":formattedDate,
-      "timecancel":formattedTime,
-      "who":'worker',
     });
     // ignore: deprecated_member_use
     return json.decode(ressponse.body);
