+++ lib/screens/signuser_screen.dart
@@ -115,7 +115,7 @@ class _Body extends State<SignuserScreen> {
   OverlayEntry _createFloatingDropdown() {
     return OverlayEntry(builder: (context) {
       return Positioned(
-        left: 50,
+        left: 55,
         top: 350,
         height: 500,
         child: Drop(
@@ -140,22 +140,8 @@ class _Body extends State<SignuserScreen> {
     String pass="";
     return Directionality(
       textDirection: TextDirection.rtl,
-      child: Stack(
-      children:[
-    //   Container(
-    //   height: 150,
-    //   decoration: BoxDecoration(
-    //   color: Colors.grey.withOpacity(0.3),
-    //   image: new DecorationImage(
-    //     fit: BoxFit.cover,
-    //     colorFilter:
-    //     ColorFilter.mode(Colors.blue.withOpacity(0.3),
-    //         BlendMode.dstATop),
-    //     image: new AssetImage('assets/work/cv.jpg',),
-    //   ),
-    // )),
-      Scaffold(
-        backgroundColor: Colors.transparent,
+      child:Scaffold(
+        backgroundColor:Colors.white,
         body: Form(key: formKey,
           child: SingleChildScrollView(
             child: Column(
@@ -172,11 +158,11 @@ class _Body extends State<SignuserScreen> {
                             height: 230,
                             width: 500,
                             decoration: BoxDecoration(
-                              color: Colors.grey.withOpacity(0.9),
+                              color: Colors.black87.withOpacity(0.9),
                               image: new DecorationImage(
                                 fit: BoxFit.cover,
                                 colorFilter:
-                                ColorFilter.mode(Colors.grey.withOpacity(0.8),
+                                ColorFilter.mode(Colors.black87.withOpacity(0.4),
                                     BlendMode.dstATop),
                                 image: new AssetImage('assets/work/cvtop.jpg',),
                               ),
@@ -224,10 +210,10 @@ class _Body extends State<SignuserScreen> {
                     child: AnimatedContainer(
                       duration: Duration(milliseconds: 500),
                       curve: Curves.ease,
-                      height: MediaQuery.of(context).size.height * 0.71,
+                      height: MediaQuery.of(context).size.height * 0.72,
                       child: Container(
-                        color: Colors.white,
-                        padding: EdgeInsets.only(top: 0),
+                        color: Colors.transparent,
+                        padding: EdgeInsets.only(top: 5),
                         child: SingleChildScrollView(
                           child: Padding(
                             padding: EdgeInsets.symmetric(horizontal: 0,),
@@ -257,6 +243,12 @@ class _Body extends State<SignuserScreen> {
                                                    return null;
                                                  }
                                                },
+                                               onChanged: (content) {
+                                                 Namelast_Null=true;
+                                                 setState(() {
+
+                                                 });
+                                               },
                                                cursorColor: kPrimaryColor,
                                                textAlign: TextAlign.right,
                                                style: TextStyle(
@@ -266,15 +258,15 @@ class _Body extends State<SignuserScreen> {
                                                ),
                                                decoration: InputDecoration(
                                                  filled: true,
-                                                 fillColor:Colors.grey[100],
+                                                 fillColor: Colors.black87.withOpacity(0.03),
                                                  enabledBorder: new OutlineInputBorder(
                                                    borderRadius: new BorderRadius.circular(30.0),
-                                                   borderSide:  BorderSide(color:Colors.grey[100]),
+                                                   borderSide:  BorderSide(color:Colors.grey[50]),
 
                                                  ),
                                                  focusedBorder: new OutlineInputBorder(
                                                    borderRadius: new BorderRadius.circular(30.0),
-                                                   borderSide:  BorderSide(color:Colors.grey[100]),
+                                                   borderSide:  BorderSide(color:Colors.grey[50]),
 
                                                  ),
                                                  hintText: 'الاسم الأول ',
@@ -304,6 +296,12 @@ class _Body extends State<SignuserScreen> {
                                                    return null;
                                                  }
                                                },
+                                               onChanged: (content) {
+                                                 Namelast_Null=true;
+                                                 setState(() {
+
+                                                 });
+                                               },
                                                cursorColor: kPrimaryColor,
                                                textAlign: TextAlign.right,
                                                style: TextStyle(
@@ -313,15 +311,15 @@ class _Body extends State<SignuserScreen> {
                                                ),
                                                decoration: InputDecoration(
                                                  filled: true,
-                                                 fillColor:Colors.grey[100],
+                                                 fillColor: Colors.black87.withOpacity(0.03),
                                                  enabledBorder: new OutlineInputBorder(
                                                    borderRadius: new BorderRadius.circular(30.0),
-                                                   borderSide:  BorderSide(color:Colors.grey[100]),
+                                                   borderSide:  BorderSide(color:Colors.grey[50]),
 
                                                  ),
                                                  focusedBorder: new OutlineInputBorder(
                                                    borderRadius: new BorderRadius.circular(30.0),
-                                                   borderSide:  BorderSide(color:Colors.grey[100]),
+                                                   borderSide:  BorderSide(color:Colors.grey[50]),
 
                                                  ),
                                                  //contentPadding: EdgeInsets.only(),
@@ -375,6 +373,13 @@ class _Body extends State<SignuserScreen> {
                                             return null;
                                           }
                                         },
+                                        onChanged: (content) {
+                                          Name_Null=true;
+                                          setState(() {
+
+                                          });
+                                        },
+
                                         cursorColor: kPrimaryColor,
                                         textAlign: TextAlign.right,
                                         style: TextStyle(
@@ -384,16 +389,16 @@ class _Body extends State<SignuserScreen> {
                                         ),
                                         decoration: InputDecoration(
                                           filled: true,
-                                          fillColor: Colors.grey[100],
-                                          prefixIcon: Icon(Icons.person,size:25,color: Color(0xFF666360)),
+                                          fillColor: Colors.black87.withOpacity(0.03),
+                                          prefixIcon: Icon(Icons.person,size:24,color: Color(0xFF666360)),
                                           enabledBorder: new OutlineInputBorder(
                                             borderRadius: new BorderRadius.circular(30.0),
-                                            borderSide:  BorderSide(color:Colors.grey[100]),
+                                            borderSide:  BorderSide(color:Colors.grey[50]),
 
                                           ),
                                           focusedBorder: new OutlineInputBorder(
                                             borderRadius: new BorderRadius.circular(30.0),
-                                            borderSide:  BorderSide(color:Colors.grey[100]),
+                                            borderSide:  BorderSide(color:Colors.grey[50]),
 
                                           ),
                                           hintText: 'اسم المستخدم ',
@@ -447,6 +452,13 @@ class _Body extends State<SignuserScreen> {
                                             Pass_S=true;
                                             return null;
                                           },
+                                          onChanged: (content) {
+                                            Pass_Null=true;
+                                            Pass_S=true;
+                                            setState(() {
+
+                                            });
+                                          },
                                           obscureText: !_showPassword1,
                                           cursorColor: kPrimaryColor,
                                           textAlign: TextAlign.right,
@@ -457,24 +469,24 @@ class _Body extends State<SignuserScreen> {
                                           ),
                                           decoration: InputDecoration(
                                             filled: true,
-                                            fillColor: Colors.grey[100],
-                                            prefixIcon: Icon(Icons.lock,size:25, color:Color(0xFF666360)),
+                                            fillColor: Colors.black87.withOpacity(0.03),
+                                            prefixIcon: Icon(Icons.lock,size:24, color:Color(0xFF666360)),
                                             suffixIcon: GestureDetector(
                                               onTap: () {
                                                 _togglevisibility1();
                                               },
                                               child: Icon(
                                                 _showPassword1 ? Icons.visibility : Icons
-                                                    .visibility_off,size:25, color: Color(0xFF666360),),
+                                                    .visibility_off,size:24, color: Color(0xFF666360),),
                                             ),
                                             enabledBorder: new OutlineInputBorder(
                                               borderRadius: new BorderRadius.circular(30.0),
-                                              borderSide:  BorderSide(color:Colors.grey[100]),
+                                              borderSide:  BorderSide(color:Colors.grey[50]),
 
                                             ),
                                             focusedBorder: new OutlineInputBorder(
                                               borderRadius: new BorderRadius.circular(30.0),
-                                              borderSide:  BorderSide(color:Colors.grey[100]),
+                                              borderSide:  BorderSide(color:Colors.grey[50]),
 
                                             ),
                                             hintText: ('كلمة السر'),
@@ -550,7 +562,7 @@ class _Body extends State<SignuserScreen> {
                                           ),
                                           decoration: InputDecoration(
                                             filled: true,
-                                            fillColor: Colors.grey[100],
+                                            fillColor: Colors.black87.withOpacity(0.03),
                                             prefixIcon: Icon(Icons.lock,size:25, color: Color(0xFF666360)),
                                             suffixIcon: GestureDetector(
                                               onTap: () {
@@ -562,12 +574,12 @@ class _Body extends State<SignuserScreen> {
                                             ),
                                             enabledBorder: new OutlineInputBorder(
                                               borderRadius: new BorderRadius.circular(30.0),
-                                              borderSide:  BorderSide(color:Colors.grey[100]),
+                                              borderSide:  BorderSide(color:Colors.grey[50]),
 
                                             ),
                                             focusedBorder: new OutlineInputBorder(
                                               borderRadius: new BorderRadius.circular(30.0),
-                                              borderSide:  BorderSide(color:Colors.grey[100]),
+                                              borderSide:  BorderSide(color:Colors.grey[50]),
 
                                             ),
                                             hintText: ('تأكيد كلمة السر'),
@@ -661,21 +673,24 @@ class _Body extends State<SignuserScreen> {
                                           width:size.width*0.8,
                                           decoration: BoxDecoration(
                                             borderRadius: BorderRadius.circular(30),
-                                            color: Colors.grey[100],
+                                            color:Colors.black87.withOpacity(0.03),
                                           ),
-                                          padding: const EdgeInsets.symmetric(horizontal: 25, vertical: 8),
+                                          padding: const EdgeInsets.only(right: 25, top: 8,bottom:8),
                                           child: Row(
                                             children: <Widget>[
-                                              Text(
-                                                country_id,
-                                                style: TextStyle(
-                                                  fontSize: 15.0,
-                                                  fontFamily: 'Changa',
-                                                  color: Color(0xFF666360),
-                                                  fontWeight: FontWeight.bold,
+                                              Container(
+                                                width: 100,
+                                                child:Text(
+                                                  country_id,
+                                                  style: TextStyle(
+                                                    fontSize: 15.0,
+                                                    fontFamily: 'Changa',
+                                                    color: Color(0xFF666360),
+                                                    fontWeight: FontWeight.bold,
+                                                  ),
                                                 ),
                                               ),
-                                              Spacer(),
+                                              SizedBox(width: 168,),
                                               Icon(
                                                 Icons.location_on,
                                                 color: Color(0xFF666360),
@@ -692,7 +707,7 @@ class _Body extends State<SignuserScreen> {
                                       //     width: size.width * 0.8,
                                       //     height: 60,
                                       //     decoration: BoxDecoration(
-                                      //       color: Colors.grey[100],
+                                      //       color: Colors.grey[50],
                                       //       borderRadius: BorderRadius.circular(30.0),
                                       //       // border: Border.all(color: Colors.grey, width: 1.2)
                                       //     ),
@@ -758,12 +773,14 @@ class _Body extends State<SignuserScreen> {
 
                                           ),),
                                       ),
-                                      Container(
+                                       Directionality(textDirection: TextDirection.rtl,
+                                      child:Container(
                                           margin: EdgeInsets.fromLTRB(0,0,0,0),
                                           padding: EdgeInsets.symmetric(),
                                           width: size.width * 0.8,
                                           height: 60,
                                           child: TextFormField(
+                                            textDirection: TextDirection.ltr,
                                             controller: phone_Num,
                                             keyboardType: TextInputType.phone,
                                             validator: (value) {
@@ -778,15 +795,15 @@ class _Body extends State<SignuserScreen> {
                                             },
                                             decoration: InputDecoration(
                                               filled: true,
-                                              fillColor: Colors.grey[100],
+                                              fillColor: Colors.black87.withOpacity(0.03),
                                               prefixIcon: Icon(Icons.phone,size:25,color: Color(0xFF666360),),
                                               enabledBorder: new OutlineInputBorder(
                                                 borderRadius: new BorderRadius.circular(30.0),
-                                                borderSide:  BorderSide(color:Colors.grey[100]),
+                                                borderSide:  BorderSide(color:Colors.grey[50]),
                                               ),
                                               focusedBorder: new OutlineInputBorder(
                                                 borderRadius: new BorderRadius.circular(30.0),
-                                                borderSide:  BorderSide(color:Colors.grey[100]),
+                                                borderSide:  BorderSide(color:Colors.grey[50]),
 
                                               ),
                                               hintText: ('رقم الهاتف'),
@@ -810,7 +827,7 @@ class _Body extends State<SignuserScreen> {
                                               fontFamily: 'Changa',
 
                                             ),
-                                          )),
+                                          )),),
                                       Phone ? Container(
                                           child: Text('',textAlign:TextAlign.end,
                                             style: TextStyle(
@@ -831,7 +848,8 @@ class _Body extends State<SignuserScreen> {
 
                                           ),),
                                       ),
-                                      Invaled_Phone ? Container(): Container(
+                                      Invaled_Phone ? Container():
+                                      Container(
                                         margin: EdgeInsets.fromLTRB(1, 0,0, 0),
                                         child: Text('يجب أن يكون رقم الهاتف على الصورة [number][country] + ',textAlign:TextAlign.end,
                                           style: TextStyle(
@@ -842,13 +860,19 @@ class _Body extends State<SignuserScreen> {
 
                                           ),),
                                       ),
-                                      invalid_OTP?Container(
-                                        margin: EdgeInsets.only(top: 10),
-                                        child: ClipRRect(
-                                          borderRadius: BorderRadius.circular(29),
-                                          child: FlatButton(
-                                              padding: EdgeInsets.symmetric(vertical: 9, horizontal: 50),
-                                              color: Color(0xFF1C1C1C),
+                                      invalid_OTP?GestureDetector(
+                                        onTap: (){
+                                          if (formKey.currentState.validate()) {print('validate');}
+                                          else{print('not validate');}
+                                          if(Phone){
+                                            _verifyPhone(phoneNo);
+                                            setState(() {
+
+                                            });
+                                          }
+                                        },
+                                        child: Container(
+                                            margin: EdgeInsets.only(top: 25),
                                               child:Text("إعادة إرسال الكود",
                                                 style:TextStyle(
                                                   fontSize: 18.0,
@@ -856,25 +880,20 @@ class _Body extends State<SignuserScreen> {
                                                   color: Color(0xFF666360),
                                                   fontWeight: FontWeight.bold,
                                                 ),),
-                                              onPressed: (){
-                                                if (formKey.currentState.validate()) {print('validate');}
-                                                else{print('not validate');}
-                                                if(Phone){
-                                                  _verifyPhone(phoneNo);
-                                                  setState(() {
-
-                                                  });
-                                                }
-                                              }
                                           ),
-                                        ),
-                                      ):Container(
-                                        margin: EdgeInsets.only(top: 10),
-                                        child: ClipRRect(
-                                          borderRadius: BorderRadius.circular(29),
-                                          child: FlatButton(
-                                              padding: EdgeInsets.symmetric(vertical:0, horizontal:0),
-                                              color: Colors.white,
+                                        ):
+                                      GestureDetector(
+                                        onTap: (){
+                                          if (formKey.currentState.validate()) {print('validate');}
+                                          else{print('not validate');}
+                                          if(Phone){
+                                            _verifyPhone(phoneNo);
+                                            setState(() {
+                                            });
+                                          }
+                                        },
+                                        child: Container(
+                                          margin: EdgeInsets.only(top: 25),
                                               child:Text("اضغط لإرسال الكود",
                                                 style:TextStyle(
                                                   fontSize: 18.0,
@@ -882,35 +901,8 @@ class _Body extends State<SignuserScreen> {
                                                   color: Color(0xFF666360),
                                                   fontWeight: FontWeight.bold,
                                                 ),),
-                                              onPressed: (){
-                                                if (formKey.currentState.validate()) {print('validate');}
-                                                else{print('not validate');}
-                                                if(Phone){
-                                                  _verifyPhone(phoneNo);
-                                                  setState(() {
-                                                  });
-                                                }
-                                                //   if(Phone){
-                                                //   verifyPhone(phoneNo);
-                                                //   setState(() {
-                                                //
-                                                //   });
-                                                // }
-                                              }
                                           ),
                                         ),
-                                      ),
-                                      // Container(
-                                      //   child: RaisedButton(
-                                      //     //child:Text("jH;d] الكود"),
-                                      //     onPressed: (){
-                                      //       if(codeSent){
-                                      //         print("send");
-                                      //         signInWithOTP(smsCode, verificationId);
-                                      //       }
-                                      //     },
-                                      //   ),
-                                      // ),
                                       codeSent?Center(
                                         child:Directionality(textDirection: TextDirection.ltr,
                                         child:Container(
@@ -932,7 +924,7 @@ class _Body extends State<SignuserScreen> {
                                               );
                                             },
                                           ), //
-                                        ),),):Container(height: 145,),
+                                        ),),):Container(height: 135,),
                                     ],
                                   ),
                                 ),
@@ -945,7 +937,7 @@ class _Body extends State<SignuserScreen> {
                                         side: BorderSide(color: Colors.transparent)
                                     ),
                                     padding: EdgeInsets.symmetric(vertical: 10, horizontal: 3),
-                                    color: Colors.black87,
+                                    color: Y,
                                     onPressed: () async {
                                       if(!card1) {
                                         setState(() {
@@ -1046,7 +1038,7 @@ class _Body extends State<SignuserScreen> {
                       ),
                     )
                 ), ],),),
-        ),),],),);
+        ),),);
   }
   _verifyPhone(String phone) async {
     await FirebaseAuth.instance.verifyPhoneNumber(
@@ -1136,39 +1128,39 @@ class _Body extends State<SignuserScreen> {
     return Column(
       children: <Widget>[
         SizedBox(
-          height: 5,
+          height: 13,
         ),
-        Container(
-          width: 300,
-         // color: Colors.yellow,
-          padding: EdgeInsets.only(right: 238),
-          alignment: Alignment.topLeft,
-        child:Align(
-         // alignment: Alignment(10,20),
-          child: ClipPath(
-            clipper: ArrowClipper(),
-            child:Container(
-              height: 20,
-              width: 30,
-              decoration: BoxDecoration(
-                  color: Colors.white,
-                boxShadow: [
-                  BoxShadow(
-                    offset: Offset(0, 1),
-                    blurRadius: 0.02,
-                    color: Colors.black,
-                  ),
-                ],
-                  // boxShadow:
-                ),
-            // child: Card(
-            //   elevation: 5,
-            //   //margin: EdgeInsets.only(left: 50,right: 50),
-            //   // height: 20,
-            //   // width: 30,
-            ),
-          ),
-        ),),
+        // Container(
+        //   width: 300,
+        //  // color: Colors.yellow,
+        //   padding: EdgeInsets.only(right: 238),
+        //   alignment: Alignment.topLeft,
+        // child:Align(
+        //  // alignment: Alignment(10,20),
+        //   child: ClipPath(
+        //     clipper: ArrowClipper(),
+        //     child:Container(
+        //       height: 20,
+        //       width: 30,
+        //       decoration: BoxDecoration(
+        //           color: Colors.white,
+        //         boxShadow: [
+        //           BoxShadow(
+        //             offset: Offset(0, 1),
+        //             blurRadius: 0.02,
+        //             color: Colors.black,
+        //           ),
+        //         ],
+        //           // boxShadow:
+        //         ),
+        //     // child: Card(
+        //     //   elevation: 5,
+        //     //   //margin: EdgeInsets.only(left: 50,right: 50),
+        //     //   // height: 20,
+        //     //   // width: 30,
+        //     ),
+        //   ),
+        // ),),
         Material(
           elevation: 5,
           //shape: ArrowShape(),
@@ -1203,16 +1195,16 @@ class _Body extends State<SignuserScreen> {
       child:Stack (children: <Widget>[
         CircleAvatar(
           backgroundImage: image_file==null? AssetImage('assets/icons/signup.jpg'):FileImage(File(image_file.path)),
-          radius: 45.0,
+          radius: 38.0,
         ),
         Positioned(
           bottom:3.0,
-          right:-1.0,
+          right:-3.0,
           child: InkWell(
             onTap:(){
               showModalBottomSheet(context: context, builder: (builder) => buttom_camera(),);
             },
-            child:Icon(Icons.camera_alt,color: Colors.black.withOpacity(0.5),size: 35.0,),),),
+            child:Icon(Icons.camera_alt,color: Colors.black.withOpacity(0.9),size: 35.0,),),),
       ],
       ),);
   }
