+++ lib/screens/signworker_screen.dart
@@ -117,7 +117,7 @@ class _Body extends State<SignWorker> {
   OverlayEntry _createFloatingDropdown() {
     return OverlayEntry(builder: (context) {
       return Positioned(
-        left: 50,
+        left: 55,
         top: 350,
         height: 500,
         child: Drop_Work(
@@ -129,7 +129,7 @@ class _Body extends State<SignWorker> {
   OverlayEntry _createFloatingDropdown_coutry() {
     return OverlayEntry(builder: (context) {
       return Positioned(
-        left: 50,
+        left: 55,
         top: 350,
         height: 500,
         child: Drop_country(
@@ -195,11 +195,11 @@ class _Body extends State<SignWorker> {
                             height: 230,
                             width: 500,
                             decoration: BoxDecoration(
-                              color: Colors.grey.withOpacity(0.9),
+                              color: Colors.black87.withOpacity(0.9),
                               image: new DecorationImage(
                                 fit: BoxFit.cover,
                                 colorFilter:
-                                ColorFilter.mode(Colors.grey.withOpacity(0.8),
+                                ColorFilter.mode(Colors.black87.withOpacity(0.4),
                                     BlendMode.dstATop),
                                 image: new AssetImage('assets/work/cvtop.jpg',),
                               ),
@@ -264,6 +264,12 @@ class _Body extends State<SignWorker> {
                                                     return null;
                                                   }
                                                 },
+                                                onChanged: (content) {
+                                                  Namefirst_Null=true;
+                                                  setState(() {
+
+                                                  });
+                                                },
                                                 cursorColor: kPrimaryColor,
                                                 textAlign: TextAlign.right,
                                                 style: TextStyle(
@@ -273,15 +279,15 @@ class _Body extends State<SignWorker> {
                                                 ),
                                                 decoration: InputDecoration(
                                                   filled: true,
-                                                  fillColor: Colors.grey[100],
+                                                  fillColor: Colors.black87.withOpacity(0.03),
                                                   enabledBorder: new OutlineInputBorder(
                                                     borderRadius: new BorderRadius.circular(30.0),
-                                                    borderSide:  BorderSide(color:Colors.grey[100]),
+                                                    borderSide:  BorderSide(color:Colors.grey[50]),
 
                                                   ),
                                                   focusedBorder: new OutlineInputBorder(
                                                     borderRadius: new BorderRadius.circular(30.0),
-                                                    borderSide:  BorderSide(color:Colors.grey[100]),
+                                                    borderSide:  BorderSide(color:Colors.grey[50]),
 
                                                   ),
                                                   hintText: 'الاسم الأول ',
@@ -311,6 +317,12 @@ class _Body extends State<SignWorker> {
                                                     return null;
                                                   }
                                                 },
+                                                onChanged: (content) {
+                                                  Namelast_Null=true;
+                                                  setState(() {
+
+                                                  });
+                                                },
                                                 cursorColor: kPrimaryColor,
                                                 textAlign: TextAlign.right,
                                                 style: TextStyle(
@@ -320,15 +332,15 @@ class _Body extends State<SignWorker> {
                                                 ),
                                                 decoration: InputDecoration(
                                                   filled: true,
-                                                  fillColor: Colors.grey[100],
+                                                  fillColor:Colors.black87.withOpacity(0.03),
                                                   enabledBorder: new OutlineInputBorder(
                                                     borderRadius: new BorderRadius.circular(30.0),
-                                                    borderSide:  BorderSide(color:Colors.grey[100]),
+                                                    borderSide:  BorderSide(color:Colors.grey[50]),
 
                                                   ),
                                                   focusedBorder: new OutlineInputBorder(
                                                     borderRadius: new BorderRadius.circular(30.0),
-                                                    borderSide:  BorderSide(color:Colors.grey[100]),
+                                                    borderSide:  BorderSide(color:Colors.grey[50]),
 
                                                   ),
                                                   //contentPadding: EdgeInsets.only(),
@@ -382,6 +394,12 @@ class _Body extends State<SignWorker> {
                                             return null;
                                           }
                                         },
+                                        onChanged: (content) {
+                                          Name_Null=true;
+                                          setState(() {
+
+                                          });
+                                        },
                                         cursorColor: kPrimaryColor,
                                         textAlign: TextAlign.right,
                                         style: TextStyle(
@@ -391,19 +409,19 @@ class _Body extends State<SignWorker> {
                                         ),
                                         decoration: InputDecoration(
                                           filled: true,
-                                          fillColor: Colors.grey[100],
-                                          prefixIcon: Icon(Icons.person,size:25,color: Color(0xFF666360)),
+                                          fillColor: Colors.black87.withOpacity(0.03),
+                                          prefixIcon: Icon(Icons.person,size:24,color: Color(0xFF666360)),
                                           enabledBorder: new OutlineInputBorder(
                                             borderRadius: new BorderRadius.circular(30.0),
-                                            borderSide:  BorderSide(color:Colors.grey[100]),
+                                            borderSide:  BorderSide(color:Colors.grey[50]),
 
                                           ),
                                           focusedBorder: new OutlineInputBorder(
                                             borderRadius: new BorderRadius.circular(30.0),
-                                            borderSide:  BorderSide(color:Colors.grey[100]),
+                                            borderSide:  BorderSide(color:Colors.grey[50]),
 
                                           ),
-                                          hintText: 'الاسم ',
+                                          hintText: 'اسم المستخدم ',
                                           hintStyle: TextStyle(
                                             fontSize: 15.0,
                                             fontFamily: 'Changa',
@@ -454,6 +472,13 @@ class _Body extends State<SignWorker> {
                                             Pass_S=true;
                                             return null;
                                           },
+                                          onChanged: (content) {
+                                            Pass_Null=true;
+                                            Pass_S=true;
+                                            setState(() {
+
+                                            });
+                                          },
                                           obscureText: !_showPassword1,
                                           cursorColor: kPrimaryColor,
                                           textAlign: TextAlign.right,
@@ -464,24 +489,24 @@ class _Body extends State<SignWorker> {
                                           ),
                                           decoration: InputDecoration(
                                             filled: true,
-                                            fillColor: Colors.grey[100],
-                                            prefixIcon: Icon(Icons.lock,size:25,color:Color(0xFF666360)),
+                                            fillColor: Colors.black87.withOpacity(0.03),
+                                            prefixIcon: Icon(Icons.lock,size:24,color:Color(0xFF666360)),
                                             suffixIcon: GestureDetector(
                                               onTap: () {
                                                 _togglevisibility1();
                                               },
                                               child: Icon(
                                                 _showPassword1 ? Icons.visibility : Icons
-                                                    .visibility_off,size:25,color: Color(0xFF666360),),
+                                                    .visibility_off,size:24,color: Color(0xFF666360),),
                                             ),
                                             enabledBorder: new OutlineInputBorder(
                                               borderRadius: new BorderRadius.circular(30.0),
-                                              borderSide:  BorderSide(color:Colors.grey[100]),
+                                              borderSide:  BorderSide(color:Colors.grey[50]),
 
                                             ),
                                             focusedBorder: new OutlineInputBorder(
                                               borderRadius: new BorderRadius.circular(30.0),
-                                              borderSide:  BorderSide(color:Colors.grey[100]),
+                                              borderSide:  BorderSide(color:Colors.grey[50]),
 
                                             ),
                                             hintText: ('كلمة السر'),
@@ -554,24 +579,24 @@ class _Body extends State<SignWorker> {
                                           ),
                                           decoration: InputDecoration(
                                             filled: true,
-                                            fillColor: Colors.grey[100],
-                                            prefixIcon: Icon(Icons.lock, size:25,color: Color(0xFF666360)),
+                                            fillColor: Colors.black87.withOpacity(0.03),
+                                            prefixIcon: Icon(Icons.lock, size:24,color: Color(0xFF666360)),
                                             suffixIcon: GestureDetector(
                                               onTap: () {
                                                 _togglevisibility2();
                                               },
                                               child: Icon(
                                                   _showPassword2 ? Icons.visibility : Icons
-                                                      .visibility_off, size:25,color: Color(0xFF666360)),
+                                                      .visibility_off, size:24,color: Color(0xFF666360)),
                                             ),
                                             enabledBorder: new OutlineInputBorder(
                                               borderRadius: new BorderRadius.circular(30.0),
-                                              borderSide:  BorderSide(color:Colors.grey[100]),
+                                              borderSide:  BorderSide(color:Colors.grey[50]),
 
                                             ),
                                             focusedBorder: new OutlineInputBorder(
                                               borderRadius: new BorderRadius.circular(30.0),
-                                              borderSide:  BorderSide(color:Colors.grey[100]),
+                                              borderSide:  BorderSide(color:Colors.grey[50]),
 
                                             ),
                                             hintText: ('تأكيد كلمة السر'),
@@ -641,47 +666,53 @@ class _Body extends State<SignWorker> {
                                             color:Colors.yellow,
                                             size: 20.0,
                                           ),),),
-                                      Container(
-                                        margin: EdgeInsets.fromLTRB(0,10,0,0),
-                                        width: size.width * 0.8,
-                                        height: 60,
-                                        child: TextFormField(
-                                          controller: Work,
-                                          validator: (value) {
-                                            if (value.isEmpty) {
-                                              Work_Null=false;
-                                              return null;
-                                            } else {
-                                              Work_Null=true;
-                                              return null;
-                                            }
-                                          },
-                                          cursorColor: Colors.grey[600],
-                                          textAlign: TextAlign.right,
-                                          style: TextStyle(
-                                            fontSize: 15.0,
-                                            fontWeight: FontWeight.bold,
-                                            fontFamily: 'Changa',
-                                          ),
-                                          decoration: InputDecoration(
-                                            filled: true,
-                                            fillColor: Colors.grey[100],
-                                            enabledBorder: new OutlineInputBorder(
-                                              borderRadius: new BorderRadius.circular(28.0),
-                                              borderSide:  BorderSide(color:Colors.grey[100]),
-
-                                            ),
-                                            focusedBorder: new OutlineInputBorder(
-                                              borderRadius: new BorderRadius.circular(28.0),
-                                              borderSide:  BorderSide(color:Colors.grey[100]),
-
-                                            ),
-                                            labelText: ('المهنة'),
-                                            floatingLabelBehavior: FloatingLabelBehavior.always,
-                                            labelStyle: TextStyle(fontSize: 22.0,fontFamily: 'Changa',color: myFocusNode.hasFocus ? Colors.grey[600] : Colors.grey[600],),
-                                          ),
-                                        ),
-                                      ),
+                                      // Container(
+                                      //   margin: EdgeInsets.fromLTRB(0,10,0,0),
+                                      //   width: size.width * 0.8,
+                                      //   height: 60,
+                                      //   child: TextFormField(
+                                      //     controller: Work,
+                                      //     validator: (value) {
+                                      //       if (value.isEmpty) {
+                                      //         Work_Null=false;
+                                      //         return null;
+                                      //       } else {
+                                      //         Work_Null=true;
+                                      //         return null;
+                                      //       }
+                                      //     },
+                                      //     onChanged: (content) {
+                                      //       Work_Null=true;
+                                      //       setState(() {
+                                      //
+                                      //       });
+                                      //     },
+                                      //     cursorColor: Colors.grey[600],
+                                      //     textAlign: TextAlign.right,
+                                      //     style: TextStyle(
+                                      //       fontSize: 15.0,
+                                      //       fontWeight: FontWeight.bold,
+                                      //       fontFamily: 'Changa',
+                                      //     ),
+                                      //     decoration: InputDecoration(
+                                      //       filled: true,
+                                      //       fillColor: Colors.black87.withOpacity(0.03),
+                                      //       enabledBorder: new OutlineInputBorder(
+                                      //         borderRadius: new BorderRadius.circular(28.0),
+                                      //         borderSide:  BorderSide(color:Colors.grey[50]),
+                                      //
+                                      //       ),
+                                      //       focusedBorder: new OutlineInputBorder(
+                                      //         borderRadius: new BorderRadius.circular(28.0),
+                                      //         borderSide:  BorderSide(color:Colors.grey[50]),
+                                      //
+                                      //       ),
+                                      //       labelText: ('المهنة'),
+                                      //       floatingLabelBehavior: FloatingLabelBehavior.always,
+                                      //       labelStyle: TextStyle(fontSize: 22.0,fontFamily: 'Changa',color: myFocusNode.hasFocus ? Colors.grey[600] : Colors.grey[600],),
+                                      //     ),
+                                      //   ),
+                                      // ),
                                       GestureDetector(
                                         onTap: () {
                                           setState(() {
@@ -700,7 +731,7 @@ class _Body extends State<SignWorker> {
                                           width:size.width*0.8,
                                           decoration: BoxDecoration(
                                             borderRadius: BorderRadius.circular(30),
-                                            color: Colors.grey[100],
+                                            color: Colors.black87.withOpacity(0.03),
                                           ),
                                           padding: const EdgeInsets.symmetric(horizontal: 25, vertical: 8),
                                           child: Row(
@@ -759,6 +790,12 @@ class _Body extends State<SignWorker> {
                                               return null;
                                             }
                                           },
+                                          onChanged: (content) {
+                                            Info_Null=true;
+                                            setState(() {
+
+                                            });
+                                          },
                                           cursorColor: Colors.grey[600],
                                           textAlign: TextAlign.right,
                                           style: TextStyle(
@@ -768,20 +805,20 @@ class _Body extends State<SignWorker> {
                                           ),
                                           decoration: InputDecoration(
                                             filled: true,
-                                            fillColor: Colors.grey[100],
+                                            fillColor: Colors.black87.withOpacity(0.03),
                                             enabledBorder: new OutlineInputBorder(
                                               borderRadius: new BorderRadius.circular(30.0),
-                                              borderSide:  BorderSide(color:Colors.grey[100]),
+                                              borderSide:  BorderSide(color:Colors.grey[50]),
 
                                             ),
                                             focusedBorder: new OutlineInputBorder(
                                               borderRadius: new BorderRadius.circular(30.0),
-                                              borderSide:  BorderSide(color:Colors.grey[100]),
+                                              borderSide:  BorderSide(color:Colors.grey[50]),
 
                                             ),
                                             labelText: ('المعلومات'),
                                             floatingLabelBehavior: FloatingLabelBehavior.always,
-                                            labelStyle: TextStyle(fontSize: 22.0,fontFamily: 'Changa',color: myFocusNode.hasFocus ? Colors.grey[600] : Colors.grey[600],),
+                                            labelStyle: TextStyle(fontSize: 18.0,fontFamily: 'Changa',color: myFocusNode.hasFocus ? Colors.grey[600] : Colors.grey[600],),
                                           ),
                                         ),
                                       ),
@@ -821,6 +858,12 @@ class _Body extends State<SignWorker> {
                                               return null;
                                             }
                                           },
+                                          onChanged: (content) {
+                                            EXp_Null=true;
+                                            setState(() {
+
+                                            });
+                                          },
                                           cursorColor: Colors.grey[600],
                                           textAlign: TextAlign.right,
                                           controller: Experiance,
@@ -831,20 +874,20 @@ class _Body extends State<SignWorker> {
                                           ),
                                           decoration: InputDecoration(
                                             filled: true,
-                                            fillColor: Colors.grey[100],
+                                            fillColor: Colors.black87.withOpacity(0.03),
                                             enabledBorder: new OutlineInputBorder(
                                               borderRadius: new BorderRadius.circular(40.0),
-                                              borderSide:  BorderSide(color:Colors.grey[100]),
+                                              borderSide:  BorderSide(color:Colors.grey[50]),
 
                                             ),
                                             focusedBorder: new OutlineInputBorder(
                                               borderRadius: new BorderRadius.circular(40.0),
-                                              borderSide:  BorderSide(color:Colors.grey[100]),
+                                              borderSide:  BorderSide(color:Colors.grey[50]),
 
                                             ),
                                             labelText: ('خبرة وتجارب سابقة'),
                                             floatingLabelBehavior: FloatingLabelBehavior.always,
-                                            labelStyle: TextStyle(fontSize: 22.0,fontFamily: 'Changa',color: myFocusNode.hasFocus ? Colors.grey[600] : Colors.grey[600],),
+                                            labelStyle: TextStyle(fontSize: 18.0,fontFamily: 'Changa',color: myFocusNode.hasFocus ? Colors.grey[600] : Colors.grey[600],),
                                           ),
                                         ),
                                       ),
@@ -868,7 +911,7 @@ class _Body extends State<SignWorker> {
                                             color: Colors.red,
                                           ),),
                                       ),
-                                      Container(height: 38,),
+                                      Container(height: 28,),
                                     ],
 
                                   ),
@@ -960,21 +1003,24 @@ class _Body extends State<SignWorker> {
                                           width:size.width*0.8,
                                           decoration: BoxDecoration(
                                             borderRadius: BorderRadius.circular(30),
-                                            color: Colors.grey[100],
+                                            color:Colors.black87.withOpacity(0.03),
                                           ),
-                                          padding: const EdgeInsets.symmetric(horizontal: 25, vertical: 8),
+                                          padding: const EdgeInsets.only(right: 25, top: 8,bottom:8),
                                           child: Row(
                                             children: <Widget>[
-                                              Text(
-                                                country_id,
-                                                style: TextStyle(
-                                                  fontSize: 15.0,
-                                                  fontFamily: 'Changa',
-                                                  color: Color(0xFF666360),
-                                                  fontWeight: FontWeight.bold,
+                                              Container(
+                                                width: 100,
+                                                child:Text(
+                                                  country_id,
+                                                  style: TextStyle(
+                                                    fontSize: 15.0,
+                                                    fontFamily: 'Changa',
+                                                    color: Color(0xFF666360),
+                                                    fontWeight: FontWeight.bold,
+                                                  ),
                                                 ),
                                               ),
-                                              Spacer(),
+                                              SizedBox(width: 168,),
                                               Icon(
                                                 Icons.location_on,
                                                 color: Color(0xFF666360),
@@ -1011,6 +1057,7 @@ class _Body extends State<SignWorker> {
                                           height: 60,
                                           child: TextFormField(
                                             controller: phone_Num,
+                                            textDirection: TextDirection.ltr,
                                             keyboardType: TextInputType.phone,
                                             validator: (value) {
                                               if (value.isEmpty) {
@@ -1024,16 +1071,16 @@ class _Body extends State<SignWorker> {
                                             },
                                             decoration: InputDecoration(
                                               filled: true,
-                                              fillColor: Colors.grey[100],
-                                              prefixIcon: Icon(Icons.phone,size:25,color: Color(0xFF666360),),
+                                              fillColor: Colors.black87.withOpacity(0.03),
+                                              prefixIcon: Icon(Icons.phone,size:24,color: Color(0xFF666360),),
                                               enabledBorder: new OutlineInputBorder(
                                                 borderRadius: new BorderRadius.circular(30.0),
-                                                borderSide:  BorderSide(color:Colors.grey[100]),
+                                                borderSide:  BorderSide(color:Colors.grey[50]),
 
                                               ),
                                               focusedBorder: new OutlineInputBorder(
                                                 borderRadius: new BorderRadius.circular(30.0),
-                                                borderSide:  BorderSide(color:Colors.grey[100]),
+                                                borderSide:  BorderSide(color:Colors.grey[50]),
 
                                               ),
                                               hintText: ('رقم الهاتف'),
@@ -1046,6 +1093,7 @@ class _Body extends State<SignWorker> {
                                               floatingLabelBehavior: FloatingLabelBehavior.always,
                                             ),
                                             onChanged: (val) {
+                                              Phone=true;
                                               setState(() {
                                                 this.phoneNo = val;
                                               });
@@ -1089,62 +1137,47 @@ class _Body extends State<SignWorker> {
 
                                           ),),
                                       ),
-                                      invalid_OTP?Container(
-                                        margin: EdgeInsets.only(top: 20),
-                                        child: ClipRRect(
-                                          borderRadius: BorderRadius.circular(29),
-                                          child: FlatButton(
-                                              padding: EdgeInsets.symmetric(vertical: 9, horizontal: 50),
-                                              child:Text("إعادة إرسال الكود",
-                                                style:TextStyle(
-                                                  fontSize: 18.0,
-                                                  fontFamily: 'Changa',
-                                                  color: Color(0xFF666360),
-                                                  fontWeight: FontWeight.bold,
-                                                ),),
-                                              onPressed: (){
-                                                if (formKey.currentState.validate()) {print('validate');}
-                                                else{print('not validate');}
-                                                if(Phone){
-                                                  _verifyPhone(phoneNo);
-                                                  setState(() {
+                                      invalid_OTP?GestureDetector(
+                                        onTap: (){
+                                          if (formKey.currentState.validate()) {print('validate');}
+                                          else{print('not validate');}
+                                          if(Phone){
+                                            _verifyPhone(phoneNo);
+                                            setState(() {
 
-                                                  });
-                                                }
-                                              }
-                                          ),
+                                            });
+                                          }
+                                        },
+                                        child: Container(
+                                          margin: EdgeInsets.only(top: 25),
+                                          child:Text("إعادة إرسال الكود",
+                                            style:TextStyle(
+                                              fontSize: 18.0,
+                                              fontFamily: 'Changa',
+                                              color: Color(0xFF666360),
+                                              fontWeight: FontWeight.bold,
+                                            ),),
                                         ),
-
-                                      ):Container(
-                                        margin: EdgeInsets.only(top: 20),
-                                        child: ClipRRect(
-                                          borderRadius: BorderRadius.circular(29),
-                                          child: FlatButton(
-                                              padding: EdgeInsets.symmetric(vertical:0, horizontal:0),
-                                              child:Text("اضغط لإرسال الكود",
-                                                style:TextStyle(
-                                                  fontSize: 18.0,
-                                                  fontFamily: 'Changa',
-                                                  color: Color(0xFF666360),
-                                                  fontWeight: FontWeight.bold,
-                                                ),),
-                                              onPressed: (){
-                                                if (formKey.currentState.validate()) {print('validate');}
-                                                else{print('not validate');}
-                                                if(Phone){
-                                                  _verifyPhone(phoneNo);
-                                                  setState(() {
-
-                                                  });
-                                                }
-                                                //   if(Phone){
-                                                //   verifyPhone(phoneNo);
-                                                //   setState(() {
-                                                //
-                                                //   });
-                                                // }
-                                              }
-                                          ),
+                                      ):
+                                      GestureDetector(
+                                        onTap: (){
+                                          if (formKey.currentState.validate()) {print('validate');}
+                                          else{print('not validate');}
+                                          if(Phone){
+                                            _verifyPhone(phoneNo);
+                                            setState(() {
+                                            });
+                                          }
+                                        },
+                                        child: Container(
+                                          margin: EdgeInsets.only(top: 25),
+                                          child:Text("اضغط لإرسال الكود",
+                                            style:TextStyle(
+                                              fontSize: 18.0,
+                                              fontFamily: 'Changa',
+                                              color: Color(0xFF666360),
+                                              fontWeight: FontWeight.bold,
+                                            ),),
                                         ),
                                       ),
                                       // Container(
@@ -1193,7 +1226,7 @@ class _Body extends State<SignWorker> {
                                         side: BorderSide(color: Colors.transparent)
                                     ),
                                     padding: EdgeInsets.symmetric(vertical: 10, horizontal: 3),
-                                    color: Colors.black87,
+                                    color: Y,
                                     onPressed: () async {
                                       if (formKey.currentState.validate()) {print('validate');}
                                       else{print('not validate');}
@@ -1376,39 +1409,8 @@ class _Body extends State<SignWorker> {
     return Column(
       children: <Widget>[
         SizedBox(
-          height: 5,
+          height: 10,
         ),
-        Container(
-          width: 300,
-          // color: Colors.yellow,
-          padding: EdgeInsets.only(right: 238),
-          alignment: Alignment.topLeft,
-          child:Align(
-            // alignment: Alignment(10,20),
-            child: ClipPath(
-              clipper: ArrowClipper(),
-              child:Container(
-                height: 20,
-                width: 30,
-                decoration: BoxDecoration(
-                  color: Colors.white,
-                  boxShadow: [
-                    BoxShadow(
-                      offset: Offset(0, 1),
-                      blurRadius: 0.02,
-                      color: Colors.black,
-                    ),
-                  ],
-                  // boxShadow:
-                ),
-                // child: Card(
-                //   elevation: 5,
-                //   //margin: EdgeInsets.only(left: 50,right: 50),
-                //   // height: 20,
-                //   // width: 30,
-              ),
-            ),
-          ),),
         Material(
           elevation: 5,
           //shape: ArrowShape(),
@@ -1437,7 +1439,7 @@ class _Body extends State<SignWorker> {
   }
   Widget DropDownItem(String text) {
 
-    return Directionality(textDirection: TextDirection.ltr,
+    return Directionality(textDirection: TextDirection.rtl,
       child:Container(
         width:300,
         height: 50,
@@ -1446,9 +1448,9 @@ class _Body extends State<SignWorker> {
           color:Colors.white,
         ),
         //margin: EdgeInsets.only(right: 16,),
-        padding: EdgeInsets.only(left: 50, top: 8,bottom: 8),
+        padding: EdgeInsets.only(top: 8,bottom: 8,right: 0),
         child: Container(
-          width: 100,
+          width: 200,
           margin: EdgeInsets.only(right: 5,),
           padding: EdgeInsets.only(right: 5,),
           alignment: Alignment.topRight,
@@ -1476,39 +1478,8 @@ class _Body extends State<SignWorker> {
     return Column(
       children: <Widget>[
         SizedBox(
-          height: 5,
+          height: 10,
         ),
-        Container(
-          width: 300,
-          // color: Colors.yellow,
-          padding: EdgeInsets.only(right: 238),
-          alignment: Alignment.topLeft,
-          child:Align(
-            // alignment: Alignment(10,20),
-            child: ClipPath(
-              clipper: ArrowClipper(),
-              child:Container(
-                height: 20,
-                width: 30,
-                decoration: BoxDecoration(
-                  color: Colors.white,
-                  boxShadow: [
-                    BoxShadow(
-                      offset: Offset(0, 1),
-                      blurRadius: 0.02,
-                      color: Colors.black,
-                    ),
-                  ],
-                  // boxShadow:
-                ),
-                // child: Card(
-                //   elevation: 5,
-                //   //margin: EdgeInsets.only(left: 50,right: 50),
-                //   // height: 20,
-                //   // width: 30,
-              ),
-            ),
-          ),),
         Material(
           elevation: 5,
           //shape: ArrowShape(),
@@ -1580,16 +1551,16 @@ class _Body extends State<SignWorker> {
       child:Stack (children: <Widget>[
         CircleAvatar(
           backgroundImage: image_file==null? AssetImage('assets/icons/signup.jpg'):FileImage(File(image_file.path)),
-          radius: 45.0,
+          radius: 38.0,
         ),
         Positioned(
           bottom:3.0,
-          right:-1.0,
+          right:-3.0,
           child: InkWell(
             onTap:(){
               showModalBottomSheet(context: context, builder: (builder) => buttom_camera(),);
             },
-            child:Icon(Icons.camera_alt,color: Colors.black.withOpacity(0.5),size: 35.0,),),),
+            child:Icon(Icons.camera_alt,color: Colors.black.withOpacity(0.9),size: 35.0,),),),
       ],
       ),);
   }
